** some docu for IsotopeModels
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@7840 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/IsotopeWaveletTransform.h
@@ -772,7 +772,7 @@ namespace OpenMS
 	template <typename PeakType>
 	void IsotopeWaveletTransform<PeakType>::initializeScan (const MSSpectrum<PeakType>& c_ref, const UInt c) 
 	{
-		data_length_ = c_ref.size();
+		data_length_ = (UInt) c_ref.size();
 		computeMinSpacing(c_ref);
 		Int wavelet_length=0, quarter_length=0;
